+++ src/Filter.php
@@ -1,0 +1,24 @@
+<?php
+namespace Pirsch;
+
+class Filter {
+    public $id;
+	public $from;
+	public $to;
+	public $path;
+	public $pattern;
+	public $language;
+	public $country;
+	public $referrer;
+	public $os;
+	public $browser;
+	public $platform;
+	public $screen_class;
+	public $utm_source;
+	public $utm_medium;
+	public $utm_campaign;
+	public $utm_content;
+	public $utm_term;
+	public $limit;
+	public $include_avg_time_on_page;
+}