+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0
 
 * added `source` and `utm_source` to referrers
+* added methods to read statistics
 
 ## 1.1.2